requirements.txt
openai>=1.3.5
PyPDF2>=3.0.0
pdf2image>=1.16.3
Pillow>=9.5.0
imagehash>=4.3.2
numpy>=1.24.0
google-api-python-client>=2.102.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
httplib2>=0.20.4